+++ Offwind.WakeFarmControl/NowCast/NowCastSimulationResult.cs
@@ -5,7 +5,7 @@ namespace WakeFarmControl.NowCast
     public class NowCastSimulationResult
     {
         public string Method;
-        public int[] Time;
+        public double[] Time;
         public double[] X;
         public double[][] XhmsAll;
         public int XhmsAllTimeOffset;
